+++ lib/src/data_source/mod.rs
@@ -1,6 +1,8 @@
 #[cfg(feature = "plugin")]
 pub mod connector_plugin;
 pub mod csv_file;
+#[cfg(feature = "flight")]
+pub mod flight_stream;
 pub mod infer_schema;
 pub mod json_file;
 pub mod json_rest;