+++ lib/src/server/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "flight")]
+pub mod flight;
 pub mod http;
 pub mod routes;
 pub mod signal_handler;